pyproject.toml
[tool.poetry]
name = "coding-challenge-screening"
version = "0.2.0"
description = ""
authors = ["ellamind"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.20.0"
huggingface-hub = "^0.24.5"
ipykernel = "^6.29.5"
openai = "^1.55.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
llama-cpp-python = "^0.3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
